refactor: move preferences input handling in a dedicate component
